Redesigned to support target context on navigation
